feat(migrations): Implement tenant-isolated asset inventory and vendor management
- Add migration 047 for tenant-isolated asset inventory, including tables for assets and asset events, along with necessary constraints and policies.
- Introduce asset_tags_are_valid function for validating asset tags.
- Create migration 048 for vendor management, establishing vendor-related tables and constraints, including vendor contacts, contracts, certifications, and subprocessors.
- Implement vendor_text_array_is_valid function for validating vendor-related text arrays.
- Add migration 049 for enterprise RBAC administration, enhancing user roles and role change events with tenant scope validation and immutable change history.
- Establish triggers and policies for role management and event tracking to ensure tenant isolation.

+++ internal/pkg/queue/envelope.go
@@ -148,8 +148,10 @@ func decodeEnvelope(body []byte) (Envelope, error) {
 
 func publishingForEnvelope(envelope Envelope, body []byte) amqp.Publishing {
 	headers := amqp.Table{
-		"x-schema-version": int32(envelope.SchemaVersion),
-		"x-attempt":        int32(envelope.Attempt),
+		// AMQP field tables support signed 64-bit integers. Envelope validation
+		// still constrains these values to the interoperable signed 32-bit range.
+		"x-schema-version": int64(envelope.SchemaVersion),
+		"x-attempt":        int64(envelope.Attempt),
 	}
 	if envelope.TenantID != "" {
 		headers["x-tenant-id"] = envelope.TenantID
